Fix inconsistent `Cargo.toml` and `pyproject.toml` path handling
